ConfigPropertiesProvider API (#466)

+++ driver/src/main/java/com/arangodb/internal/ArangoDefaults.java
@@ -38,7 +38,7 @@ public final class ArangoDefaults {
     public static final int CHUNK_MAX_HEADER_SIZE = CHUNK_MIN_HEADER_SIZE + LONG_BYTES;
     public static final int CHUNK_DEFAULT_CONTENT_SIZE = 30000;
     public static final int MAX_CONNECTIONS_VST_DEFAULT = 1;
-    public static final Integer CONNECTION_TTL_VST_DEFAULT = null;
+    public static final Long CONNECTION_TTL_VST_DEFAULT = null;
     public static final int MAX_CONNECTIONS_HTTP_DEFAULT = 20;
     public static final int MAX_CONNECTIONS_HTTP2_DEFAULT = 1;
     public static final Protocol DEFAULT_NETWORK_PROTOCOL = Protocol.HTTP2_JSON;
